+++ virtual_rainforest/models/soil/soil_schema.json
